refactor: refactor cart-page
