+++ src/components/promocode/messages.ts
@@ -1,4 +1,5 @@
 export const messages = {
   promoCodeTitle: 'SALE20',
   saleTitle: '20% promo',
+  totalCost: 'Total cost ',
 };
