+++ src/components/promocode/promo-code-input-panel/promo-code-input.module.css
@@ -1,0 +1,17 @@
+.promoCodeSection {
+  display: flex;
+  align-items: center;
+  border: 1px solid var(--clr-accent-11);
+  border-radius: 5px;
+  margin: 20px 0 0;
+}
+
+.promoCodeInput {
+  font-size: 16px;
+  padding-left: 6px;
+  color: var(--clr-secondary-800);
+}
+
+.promoCodeInput::placeholder {
+  font-family: var(--ff-base);
+}
