+++ src/pages/cart-page/cart-page.module.css
@@ -23,37 +23,3 @@
   align-self: flex-end;
   margin-right: 20px;
 }
-
-.promoCodeSection {
-  display: flex;
-  align-items: center;
-  border: 1px solid var(--clr-accent-11);
-  border-radius: 5px;
-  margin: 20px 0 0;
-}
-
-.promoCodeInput {
-  font-size: 16px;
-  padding-left: 6px;
-  color: var(--clr-secondary-800);
-}
-
-.promoCodeInput::placeholder {
-  font-family: var(--ff-base);
-}
-
-.priceContainer {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.priceWithoutDiscount {
-  text-decoration: line-through;
-  font-size: 20px;
-}
-
-.discountPrice {
-  color: var(--clr-accent-1);
-  font-size: 20px;
-}
